light refactoring of entities
